Move IntentConfirmParams to its own file
